+++ src/main/java/net/meander/subtlyd/mixin/common/world/entity/TamableAnimalMixin.java
@@ -1,4 +1,4 @@
-package net.meander.subtlyd.mixin.common.entity;
+package net.meander.subtlyd.mixin.common.world.entity;
 
 import net.meander.subtlyd.world.entity.TamableAnimalSD;
 import net.minecraft.world.entity.TamableAnimal;